Move PersistRepository to mode module
