+++ shardingsphere-mode/shardingsphere-mode-core/src/main/java/org/apache/shardingsphere/mode/persist/PersistRepository.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.mode.repository;
+package org.apache.shardingsphere.mode.persist;
 
 import org.apache.shardingsphere.infra.spi.typed.TypedSPI;
