+++ shardingsphere-mode/shardingsphere-mode-core/src/main/java/org/apache/shardingsphere/mode/persist/PersistService.java
@@ -20,7 +20,6 @@ package org.apache.shardingsphere.mode.persist;
 import lombok.Getter;
 import org.apache.shardingsphere.infra.config.RuleConfiguration;
 import org.apache.shardingsphere.infra.config.datasource.DataSourceConfiguration;
-import org.apache.shardingsphere.infra.mode.repository.PersistRepository;
 import org.apache.shardingsphere.mode.persist.service.SchemaMetaDataPersistService;
 import org.apache.shardingsphere.mode.persist.service.impl.DataSourcePersistService;
 import org.apache.shardingsphere.mode.persist.service.impl.GlobalRulePersistService;
